+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 ## Release Notes
 
 **Bug Fixes**
+
+- Fixed errorneous time field initialization for existing competitions.